requirements.txt
arrow==1.2.3
asgiref==3.7.2
binaryornot==0.4.4
branca==0.6.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
cookiecutter==2.2.2
Django==4.2.2
django-appconf==1.0.5
django-browser-reload==1.11.0
django-compressor==4.4
django-tailwind==3.6.0
django-webpack-loader==2.0.1
folium==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
nodeenv==1.8.0
numpy==1.25.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
PyYAML==6.0
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
